asy: introduce hw ops abstraction for UART variant support
Introduce asy_hw_ops_t, a hardware operations vector that abstracts all
16550-specific register access behind a function pointer table.  This is
preparation for supporting non-16550 UARTs (PL011/SBSA) behind the
common asy driver framework.

The ops table covers interrupt handling, chip identification, baud rate,
line control, modem control/status, line status, interrupt
enable/disable, FIFO management, receive/transmit data, status flush,
and polled console I/O.

All 16550-specific code is extracted from asy.c into the new
asy_16550.c: register accessor infrastructure, baud rate divisor table,
register table, chip identification, FIFO management, the ISR, and
polled I/O.  These are wired into the exported asy_16550_ops table.

The shared asy.c driver now calls through asy->asy_hw->aho_* for all
hardware interaction.  Functions shared between asy.c and variant
drivers (async_rxint, async_txint, async_msint, asysetsoft, asyerror,
asy_get_bus_type) are made non-static with declarations in asy.h.  A
mapfile is introduced to ensure that these symbols do not leak into
the kernel.

Introducing the mapfile helped to indentify a set of symbols leaked
into the global namespace, which are now local to the module.  These
are: asy_glob_lock, asy_soft_state, asy_ops, asy_info, asy_str_info and
async_put_suspq.  This leakage has been fixed in source as well as
implicitly via the mapfile.

No behavioral changes - this is a mechanical refactoring only.
