+++ usr/src/uts/aarch64/asy/Makefile
@@ -44,6 +44,7 @@ MODULE		= asy
 OBJECTS		= $(ASY_OBJS:%=$(OBJS_DIR)/%)
 ROOTMODULE	= $(ROOT_DRV_DIR)/$(MODULE)
 CONF_SRCDIR	= $(UTSBASE)/common/io
+MAPFILE		= $(UTSBASE)/common/io/$(MODULE).mapfile
 
 #
 #	Include common rules.
@@ -55,6 +56,8 @@ include $(UTSBASE)/aarch64/Makefile.aarch64
 #
 ALL_TARGET	= $(BINARY) $(SRC_CONFILE)
 INSTALL_TARGET	= $(BINARY) $(ROOTMODULE) $(ROOT_CONFFILE)
+
+LDFLAGS		+= -M $(MAPFILE)
 
 #
 #	Default build targets.
