+++ usr/src/uts/common/Makefile.files
@@ -969,7 +969,7 @@ KB8042_OBJS +=	\
 
 MOUSE8042_OBJS += mouse8042.o
 
-ASY_OBJS +=	asy.o
+ASY_OBJS +=	asy.o asy_16550.o
 
 ECPP_OBJS +=	ecpp.o
